+++ sdk/search/search-documents/samples/v12/javascript/vectorSearch.js
@@ -87,6 +87,14 @@ async function main() {
             kNearestNeighborsCount: 3,
             vector: luxuryQueryVector,
           },
+          // The index can be configured with a vectorizer to generate text embeddings
+          // from a text query
+          {
+            kind: "text",
+            fields: ["descriptionVectorFr"],
+            kNearestNeighborsCount: 3,
+            text: "What are the most luxurious hotels?",
+          },
         ],
       },
     });