+++ sdk/search/search-documents/test/public/node/searchIndexClient.spec.ts
@@ -1,12 +1,13 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-import { isLiveMode, Recorder } from "@azure-tools/test-recorder";
+import { env, isLiveMode, Recorder } from "@azure-tools/test-recorder";
 import { delay } from "@azure/core-util";
 import { assert } from "chai";
 import { Context, Suite } from "mocha";
 import {
   AzureKeyCredential,
+  AzureOpenAIVectorizer,
   SearchIndex,
   SearchIndexClient,
   SynonymMap,
@@ -266,33 +267,6 @@ describe("SearchIndexClient", function (this: Suite) {
         assert.equal(index.fields.length, 6);
       });
     });
-  });
-
-  describe("preview", function () {
-    let recorder: Recorder;
-    let indexClient: SearchIndexClient;
-    let TEST_INDEX_NAME: string;
-
-    beforeEach(async function (this: Context) {
-      recorder = new Recorder(this.currentTest);
-      TEST_INDEX_NAME = createRandomIndexName();
-      ({ indexClient, indexName: TEST_INDEX_NAME } = await createClients<Hotel>(
-        defaultServiceVersion,
-        recorder,
-        TEST_INDEX_NAME,
-      ));
-
-      await createSynonymMaps(indexClient);
-      await createSimpleIndex(indexClient, TEST_INDEX_NAME);
-      await delay(WAIT_TIME);
-    });
-
-    afterEach(async function () {
-      await indexClient.deleteIndex(TEST_INDEX_NAME);
-      await delay(WAIT_TIME);
-      await deleteSynonymMaps(indexClient);
-      await recorder?.stop();
-    });
 
     it("creates the index object vector fields", async function () {
       const indexName: string = isLiveMode() ? createRandomIndexName() : "hotel-live-test4";
@@ -302,10 +276,19 @@ describe("SearchIndexClient", function (this: Suite) {
         kind: "hnsw",
         parameters: { m: 10, efSearch: 1000, efConstruction: 1000, metric: "dotProduct" },
       };
-
+      const vectorizer: AzureOpenAIVectorizer = {
+        kind: "azureOpenAI",
+        vectorizerName: "vectorizer",
+        parameters: {
+          deploymentId: env.AZURE_OPENAI_DEPLOYMENT_NAME,
+          resourceUrl: env.AZURE_OPENAI_ENDPOINT,
+          modelName: "text-embedding-ada-002",
+        },
+      };
       const profile: VectorSearchProfile = {
         name: "profile",
         algorithmConfigurationName: algorithm.name,
+        vectorizerName: vectorizer.vectorizerName,
       };
 
       let index: SearchIndex = {
@@ -326,6 +309,7 @@ describe("SearchIndexClient", function (this: Suite) {
         ],
         vectorSearch: {
           algorithms: [algorithm],
+          vectorizers: [vectorizer],
           profiles: [profile],
         },
       };
@@ -333,6 +317,10 @@ describe("SearchIndexClient", function (this: Suite) {
         await indexClient.createOrUpdateIndex(index);
         index = await indexClient.getIndex(indexName);
         assert.deepEqual(index.vectorSearch?.algorithms?.[0].name, algorithm.name);
+        assert.deepEqual(
+          index.vectorSearch?.vectorizers?.[0].vectorizerName,
+          vectorizer.vectorizerName,
+        );
         assert.deepEqual(index.vectorSearch?.profiles?.[0].name, profile.name);
       } finally {
         await indexClient.deleteIndex(index);